feat(git-commit-style-guide): add general guide for git commit msgs
Git commit messages create a log of changes in a git repo. It is
important that the message is clear about what that commit does.
While we already have some basic guidelines for git messages,
this is a more structured style guide.

The advanages of adhering to a format are having logs that are
easy to read at a glance and easy to search and filter, having a
good record of the motivations and changes in a commit, and that
it makes it easier to write the commit message since you don't
have to think about the format.