update Glide documentation
First, the glide command has gotten a bit better since this was
written. Additionally, for now we're going to check in our dependencies
in new Glide repositories to make the comparison between Glide and Godep
eaiser. After that, if we settle on Glide, we can revisit whether or not
we want to check in our dependencies.