New major version of golang.mk
The primary purpose of this version bump is to support go modules. To
that end, the new file does not include anything about our old
dependency management tool, dep, or in fact anything about modules
themselves. All dependency management is now expected to be in a
repo's primary Makefile. Typically, all that is needed is a target
like this:

install_deps:
	go mod vendor
        # possibly some additional toolchain setup like go-mock

In order to facilitate the ugprade, a two-step process is
introduced. First, the previous commit added a new version of the v0.x
line of golang.mk which includes an golang-major-update-makefile
target. So a typical repository can update to this new golang-v1.mk
either directly by copying this file into the repo, or indirectly, by
updating to the latest golang.mk v0.x and running that make target
when the repository is ready to be ugpraded.

The plan going forward is for the files in this repository to be named
golang-vX.mk where X is the major version; that way we can maintain
multiple incompatible versions at once if desired.

There is one major internal difference between dependency management
with dep and with modules. With modules, there is not necessarily an
equivalent between the import path of the module and the file
path. For example, we might have a repository that you import via

`import github.com/Clever/ark/v3/cmd`

But its filepath may not include any v3 anywhere. That that end, we
are instituting the following contract for golang-v1.mk:
- the helper functions for lint, vet, fmt, test, build all
accept *IMPORT PATH* as an argument.

Of these helpers, they all can accept import paths of packages EXCEPT
gofmt, which operates on individual files. `golint` CAN operate on
packages, however there seems to be an existing mechansism by which
any individual files prefixed with `gen_` are not linted, so I have
kept that in-tact and `golint` will also operate in file-mode.

We use `go list` to convert into a file path:

`go list -f '{{ .Dir }}' some/import/path`

will output the filepath on the local system of that package. As a
result, I believe we may not need to run these make commands from
within GOPATH, however I am explicitly not testing or suggesting we
support such use cases.