lambda.mk: migrate to provided.al2 runtime
the provided.al2 runtime in lambda is the only runtime that supports
Go and lambda layers.

This runtime assumes that the executable is named `bootstrap`:
https://docs.aws.amazon.com/lambda/latest/dg/runtimes-custom.html

This PR changes our lambda build process to include the binary with
the name `bootstrap`. It preserves the old behavior for
backwards-compatability while catapult changes to deploy things with
the new runtime.